requirements.txt
pillow
numpy
tensorflow
opencv
tensorflow-gpu
requests
